#1373 * Rearranged fcp.watchfolder files * Split out the `panel` specific code.
