+++ src/plugins/finalcutpro/watchfolders/media/MediaFolder.lua
@@ -1,4 +1,4 @@
---- === plugins.finalcutpro.watchfolders.panels.media.MediaFolder ===
+--- === plugins.finalcutpro.watchfolders.media.MediaFolder ===
 ---
 --- Final Cut Pro Media Watch Folder Plugin.
 
@@ -55,7 +55,7 @@ local MediaFolder = {}
 MediaFolder.mt = {}
 MediaFolder.mt.__index = MediaFolder.mt
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder.new() -> MediaFolder
+--- plugins.finalcutpro.watchfolders.media.MediaFolder.new() -> MediaFolder
 --- Constructor
 --- Creates a new Media Folder.
 ---
@@ -83,7 +83,7 @@ function MediaFolder.new(mod, path, videoTag, audioTag, imageTag)
     }, MediaFolder.mt)
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder.thaw(details) -> MediaFolder
+--- plugins.finalcutpro.watchfolders.media.MediaFolder.thaw(details) -> MediaFolder
 --- Constructor
 --- Creates a new MediaFolder based on the details provided.
 --- The details have typically come from a call to `MediaFolder.freeze(...)`
@@ -102,7 +102,7 @@ function MediaFolder.thaw(mod, details)
     return mf
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder.freeze(mediaFolder) -> table
+--- plugins.finalcutpro.watchfolders.media.MediaFolder.freeze(mediaFolder) -> table
 --- Function
 --- Returns a table with the details of the `MediaFolder`, ready to be stored.
 --- It can be brought back via the `MediaFolder.thaw(...)` function.
@@ -123,7 +123,7 @@ function MediaFolder.freeze(mediaFolder)
 end
 MediaFolder.mt.freeze = MediaFolder.freeze
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:init() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:init() -> nil
 --- Method
 --- Initialises the folder, getting any watchers, notifications, etc. running.
 ---
@@ -163,7 +163,7 @@ function MediaFolder.mt:init()
     return self
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:doTagFiles(files) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:doTagFiles(files) -> nil
 --- Method
 --- Tags a table of files.
 ---
@@ -250,7 +250,7 @@ local function notificationDelivered(notification)
     return false
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:checkNotifications() -> none
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:checkNotifications() -> none
 --- Method
 --- Checks Notifications.
 ---
@@ -267,7 +267,7 @@ function MediaFolder.mt:checkNotifications()
     end
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:processFiles() -> none
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:processFiles() -> none
 --- Method
 --- Process files.
 ---
@@ -295,7 +295,7 @@ function MediaFolder.mt:processFiles(files, fileFlags)
     end
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:removeFile(file) -> MediaFolder
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:removeFile(file) -> MediaFolder
 --- Method
 --- Removes the file from any queues it might be in, updating relevant notifications.
 ---
@@ -317,7 +317,7 @@ function MediaFolder.mt:removeFile(file)
     return self
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:addIncoming(file) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:addIncoming(file) -> nil
 --- Method
 --- Adds the file to the 'incoming' list and updates the notification.
 ---
@@ -332,7 +332,7 @@ function MediaFolder.mt:addIncoming(file)
     self:updateIncomingNotification()
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:addReady(file) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:addReady(file) -> nil
 --- Method
 --- Adds the file to the 'ready' list and updates the notifications.
 ---
@@ -352,7 +352,7 @@ function MediaFolder.mt:addReady(file)
     end
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:updateIncomingNotification() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:updateIncomingNotification() -> nil
 --- Method
 --- Updates the 'incoming' notification based on the current set of files in the `incoming` queue.
 ---
@@ -384,7 +384,7 @@ function MediaFolder.mt:updateIncomingNotification()
     end
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:importTag() -> string
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:importTag() -> string
 --- Method
 --- Returns the import tag.
 ---
@@ -397,7 +397,7 @@ function MediaFolder.mt:importTag()
     return "import:"..self.path
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:updateReadyNotification() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:updateReadyNotification() -> nil
 --- Method
 --- Updates the 'ready' notification based on the current set of files in the `ready` queue.
 ---
@@ -452,7 +452,7 @@ function MediaFolder.mt:updateReadyNotification()
     end
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:handleImport(notification) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:handleImport(notification) -> nil
 --- Method
 --- Handles the importing of a file.
 ---
@@ -499,7 +499,7 @@ function MediaFolder.mt:handleImport(notification)
     end
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:importAll() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:importAll() -> nil
 --- Method
 --- Begins importing all `ready` files, removing them from the `ready` queue.
 ---
@@ -517,7 +517,7 @@ function MediaFolder.mt:importAll()
     self:updateReadyNotification()
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:importFirst() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:importFirst() -> nil
 --- Method
 --- Begins importing the first `ready` file, removing it from the `ready` queue.
 ---
@@ -535,7 +535,7 @@ function MediaFolder.mt:importFirst()
     self:updateReadyNotification()
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:skipAll() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:skipAll() -> nil
 --- Method
 --- Skip all files in the Media Folder.
 ---
@@ -550,7 +550,7 @@ function MediaFolder.mt:skipAll()
     self:updateReadyNotification()
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:skipOne() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:skipOne() -> nil
 --- Method
 --- Skip one file in the Media Folder.
 ---
@@ -567,7 +567,7 @@ function MediaFolder.mt:skipOne()
     end
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:importFiles(files) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:importFiles(files) -> nil
 --- Method
 --- Requests for the files to be imported.
 ---
@@ -583,7 +583,7 @@ function MediaFolder.mt:importFiles(files)
     self:doImportNext():TimeoutAfter(30000, "Import Next took too long"):Now()
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:doWriteFilesToPasteboard(files, context) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:doWriteFilesToPasteboard(files, context) -> nil
 --- Method
 --- Write files to the Pasteboard.
 ---
@@ -623,7 +623,7 @@ function MediaFolder.mt:doWriteFilesToPasteboard(files, context)
     :Label("MediaFolder:doWriteFilesToPasteboard")
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:doRestoreOriginalPasteboard(context) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:doRestoreOriginalPasteboard(context) -> nil
 --- Method
 --- Restore original Pasteboard contents after 2 seconds.
 ---
@@ -646,7 +646,7 @@ function MediaFolder.mt:doRestoreOriginalPasteboard(context)
     :Label("MediaFolder:doRestoreOriginalPasteboard")
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:doDeleteImportedFiles(context) -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:doDeleteImportedFiles(context) -> nil
 --- Method
 --- Checks if we are deleting after import, and if so schedules them to be deleted.
 ---
@@ -678,7 +678,7 @@ function MediaFolder.mt:doDeleteImportedFiles(files)
     :Label("MediaFolder:doDeleteImportedFiles")
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:doImportNext() -> nil
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:doImportNext() -> nil
 --- Method
 --- Imports the next file in the Media Folder.
 ---
@@ -769,7 +769,7 @@ function MediaFolder.mt:doImportNext()
     :Label("MediaFolder:doImportNext")
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:save()
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:save()
 --- Method
 --- Ensures the MediaFolder is saved.
 ---
@@ -782,7 +782,7 @@ function MediaFolder.mt:save()
     self.mod.saveMediaFolders()
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:destroy()
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:destroy()
 --- Method
 --- Destroys the MediaFolder. It should not be used after this is called.
 ---
@@ -814,7 +814,7 @@ function MediaFolder.mt:destroy()
     self.importing = nil
 end
 
---- plugins.finalcutpro.watchfolders.panels.media.MediaFolder:doRevealInFinder() -> cp.rx.go.Statement
+--- plugins.finalcutpro.watchfolders.media.MediaFolder:doRevealInFinder() -> cp.rx.go.Statement
 --- Method
 --- Returns a `Statement` that will reveal the MediaFolder path in the Finder.
 ---
